moves periods
