+++ lib/src/periods/period.dart
@@ -1,8 +1,8 @@
 import 'package:equatable/equatable.dart';
 import 'package:meta/meta.dart';
-import 'package:time/time.dart';
 
-import '../period.dart';
+import '../helpers/helpers.dart';
+import '../period_generator.dart';
 
 /// A period of time between two [DateTime]s.
 @immutable
@@ -678,463 +678,4 @@ class Period with EquatableMixin implements Comparable<Period> {
 
   @override
   List<Object?> get props => [start, end];
-}
-
-/// A class that implements a period type of a second.
-class SecondPeriod extends Period {
-  /// A class that implements a period type of a second.
-  SecondPeriod({required super.start, required super.end})
-      : assert(
-          end.difference(start) ==
-              const Duration(
-                milliseconds: 999,
-                microseconds: 999,
-              ),
-          'The difference between start and end must be 999 milliseconds and '
-          '999 microseconds',
-        ) {
-    const microsecond = Duration(microseconds: 1);
-    if ((duration != const Duration(seconds: 1)) ||
-        !start.isAtSameSecondAs(end) ||
-        end.add(microsecond).isAtSameSecondAs(start)) {
-      throw ArgumentError.value(
-        end,
-        'end',
-        'End must be at the same second as start and must be the last '
-            'microsecond of the second',
-      );
-    }
-  }
-}
-
-/// A class that implements a period type of a minute.
-class MinutePeriod extends Period {
-  /// A class that implements a period type of a minute.
-  MinutePeriod({required super.start, required super.end})
-      : assert(
-          end.difference(start) ==
-              const Duration(
-                seconds: 59,
-                milliseconds: 999,
-                microseconds: 999,
-              ),
-          'The difference between start and end must be 59 seconds, 999 '
-          'milliseconds and 999 microseconds',
-        ) {
-    const microsecond = Duration(microseconds: 1);
-    if (duration != const Duration(minutes: 1) ||
-        !start.isAtSameMinuteAs(end) ||
-        end.add(microsecond).isAtSameMinuteAs(start)) {
-      throw ArgumentError.value(
-        end,
-        'end',
-        'End must be at the same minute as start and must be the last '
-            'microsecond of the minute',
-      );
-    }
-  }
-
-  /// Returns the list of seconds that are contained in this minute.
-  List<SecondPeriod> get seconds {
-    const generator = SecondGenerator();
-    final seconds = <SecondPeriod>[];
-    var last = generator.of(start);
-    while (last.start.isBefore(end)) {
-      seconds.add(last);
-      last = generator.after(last);
-    }
-    return seconds;
-  }
-}
-
-/// A class that implements a period type of an hour.
-class HourPeriod extends Period {
-  /// A class that implements a period type of an hour.
-  HourPeriod({required super.start, required super.end})
-      : assert(
-          end.difference(start) ==
-              const Duration(
-                minutes: 59,
-                seconds: 59,
-                milliseconds: 999,
-                microseconds: 999,
-              ),
-          'The difference between start and end must be 59 minutes, 59 '
-          'seconds, 999 milliseconds and 999 microseconds',
-        ) {
-    const microsecond = Duration(microseconds: 1);
-    if (duration != const Duration(hours: 1) ||
-        !start.isAtSameHourAs(end) ||
-        end.add(microsecond).isAtSameHourAs(start)) {
-      throw ArgumentError.value(
-        end,
-        'end',
-        'End must be at the same hour as start and must be the last '
-            'microsecond of the hour',
-      );
-    }
-  }
-
-  /// Returns the list of minutes that are contained in this hour.
-  List<MinutePeriod> get minutes {
-    const generator = MinuteGenerator();
-    final minutes = <MinutePeriod>[];
-    var last = generator.of(start);
-    while (last.start.isBefore(end)) {
-      minutes.add(last);
-      last = generator.after(last);
-    }
-    return minutes;
-  }
-}
-
-/// A class that implements a period type of a day.
-class DayPeriod extends Period {
-  /// A class that implements a period type of a day.
-  DayPeriod({required super.start, required super.end})
-      : assert(
-          end.difference(start) < const Duration(days: 1),
-          'The difference between start and end must be 23 hours, 59 minutes, '
-          '59 seconds, 999 milliseconds and 999 microseconds',
-        ) {
-    const microsecond = Duration(microseconds: 1);
-    if (!start.isAtSameDayAs(end) ||
-        end.add(microsecond).isAtSameDayAs(start)) {
-      throw ArgumentError.value(
-        end,
-        'end',
-        'End must be at the same day as start and must be the last microsecond '
-            'of the day',
-      );
-    }
-  }
-
-  /// Returns the list of hours in this day.
-  List<HourPeriod> get hours {
-    const generator = HourGenerator();
-    final hours = <HourPeriod>[];
-    var last = generator.of(start);
-    while (last.start.isBefore(end)) {
-      hours.add(last);
-      last = generator.after(last);
-    }
-    return hours;
-  }
-}
-
-/// {@template dayPeriodBundle}
-/// A base class that represents a bundle of days.
-/// {@endtemplate}
-abstract class DayPeriodBundle extends Period {
-  /// {@macro dayPeriodBundle}
-  DayPeriodBundle({required super.start, required super.end});
-
-  /// Returns the list of days in this bundle.
-  List<DayPeriod> get days;
-}
-
-/// A class that implements a period type of a week.
-class WeekPeriod extends Period implements DayPeriodBundle {
-  /// A class that implements a period type of a week.
-  WeekPeriod({required super.start, required super.end})
-      : assert(
-          end.difference(start) <=
-              const Duration(
-                days: 7,
-                minutes: 59,
-                seconds: 59,
-                milliseconds: 999,
-                microseconds: 999,
-              ),
-          'The difference between start and end must be 7 days, 0 hours, '
-          '59 minutes, 59 seconds, 999 milliseconds and 999 microseconds',
-        ) {
-    if ((duration > const Duration(days: 7, hours: 1)) ||
-        (duration <
-            const Duration(
-              days: 6,
-              hours: 22,
-              minutes: 59,
-              seconds: 59,
-              milliseconds: 999,
-              microseconds: 999,
-            )) ||
-        (start.timeOfDay != Duration.zero) ||
-        (end.timeOfDay != end.endOfDay.timeOfDay) ||
-        (end.weekday != Weekday.from(start).previous.dateTimeValue)) {
-      throw ArgumentError.value(
-        end,
-        'end',
-        'End must be at the same week as start and must be the last '
-            'microsecond of the week',
-      );
-    }
-  }
-
-  @override
-  List<DayPeriod> get days {
-    const generator = DayGenerator();
-    final days = <DayPeriod>[];
-    var last = generator.of(start);
-    while (last.start.isBefore(end)) {
-      days.add(last);
-      last = generator.after(last);
-    }
-    return days;
-  }
-}
-
-/// A class that implements a period type of a fortnight.
-class FortnightPeriod extends Period implements DayPeriodBundle {
-  /// A class that implements a period type of a fortnight.
-  FortnightPeriod({required super.start, required super.end})
-      : assert(
-          end.difference(start) <=
-              const Duration(
-                days: 15,
-                hours: 23,
-                minutes: 59,
-                seconds: 59,
-                milliseconds: 999,
-                microseconds: 999,
-              ),
-          'The difference between start and end must be 15 days, 23 hours, '
-          '59 minutes, 59 seconds, 999 milliseconds and 999 microseconds',
-        ) {
-    if ((duration > const Duration(days: 16, hours: 1)) ||
-        (duration <
-            const Duration(
-              days: 12,
-              hours: 22,
-              minutes: 59,
-              seconds: 59,
-              milliseconds: 999,
-              microseconds: 999,
-            )) ||
-        (start.timeOfDay != Duration.zero) ||
-        (end.timeOfDay != end.endOfDay.timeOfDay) ||
-        !start.isAtSameMonthAs(end) ||
-        ((start.firstDayOfMonth != start) &&
-            (start != start.copyWith(day: 16))) ||
-        ((end.lastDayOfMonth.endOfDay != end) &&
-            (end != end.copyWith(day: 15).endOfDay))) {
-      throw ArgumentError.value(
-        end,
-        'end',
-        'End must be at the same fortnight as start and must be the last '
-            'microsecond of the fortnight',
-      );
-    }
-  }
-
-  @override
-  List<DayPeriod> get days {
-    const generator = DayGenerator();
-    final days = <DayPeriod>[];
-    var last = generator.of(start);
-    while (last.start.isBefore(end)) {
-      days.add(last);
-      last = generator.after(last);
-    }
-    return days;
-  }
-}
-
-/// A class that implements a period type of a month.
-class MonthPeriod extends Period implements DayPeriodBundle {
-  /// A class that implements a period type of a month.
-  MonthPeriod({required super.start, required super.end})
-      : assert(
-          end.difference(start) <=
-              const Duration(
-                days: 30,
-                hours: 23,
-                minutes: 59,
-                seconds: 59,
-                milliseconds: 999,
-                microseconds: 999,
-              ),
-          'The difference between start and end must be 30 days, 23 hours, '
-          '59 minutes, 59 seconds, 999 milliseconds and 999 microseconds',
-        ) {
-    const microsecond = Duration(microseconds: 1);
-    if (!start.isAtSameMonthAs(end) ||
-        end.add(microsecond).isAtSameMonthAs(start) ||
-        (start.timeOfDay != Duration.zero) ||
-        (end.timeOfDay != end.endOfDay.timeOfDay) ||
-        (start.day != 1) ||
-        (end.day != end.lastDayOfMonth.day)) {
-      throw ArgumentError.value(
-        end,
-        'end',
-        'End must be at the same month as start and must be the last '
-            'microsecond of the month',
-      );
-    }
-  }
-
-  @override
-  List<DayPeriod> get days {
-    const generator = DayGenerator();
-    final days = <DayPeriod>[];
-    var last = generator.of(start);
-    while (last.start.isBefore(end)) {
-      days.add(last);
-      last = generator.after(last);
-    }
-    return days;
-  }
-}
-
-/// {@template monthPeriodBundle}
-/// A base class that represents a bundle of months.
-/// {@endtemplate}
-abstract class MonthPeriodBundle extends Period {
-  /// {@macro monthPeriodBundle}
-  MonthPeriodBundle({required super.start, required super.end});
-
-  /// Returns the list of months in this bundle.
-  List<MonthPeriod> get months;
-}
-
-/// A class that implements a period type of a trimester.
-class TrimesterPeriod extends Period implements MonthPeriodBundle {
-  /// A class that implements a period type of a trimester.
-  TrimesterPeriod({required super.start, required super.end})
-      : assert(
-          end.difference(start) <=
-              const Duration(
-                days: 91,
-                hours: 23,
-                minutes: 59,
-                seconds: 59,
-                milliseconds: 999,
-                microseconds: 999,
-              ),
-          'The difference between start and end must be 91 days, 23 hours, '
-          '59 minutes, 59 seconds, 999 milliseconds and 999 microseconds',
-        ) {
-    if ((start.timeOfDay != Duration.zero) ||
-        (end.timeOfDay != end.endOfDay.timeOfDay) ||
-        start.isAtSameMonthAs(end) ||
-        (start.firstDayOfMonth != start) ||
-        (end.lastDayOfMonth.endOfDay != end) ||
-        ((end.month - start.month) != 2)) {
-      throw ArgumentError.value(
-        end,
-        'end',
-        'End must be at the same trimester as start and must be the last '
-            'microsecond of the trimester',
-      );
-    }
-  }
-
-  @override
-  List<MonthPeriod> get months {
-    const generator = MonthGenerator();
-    final months = <MonthPeriod>[];
-    var last = generator.of(start);
-    while (last.start.isBefore(end)) {
-      months.add(last);
-      last = generator.after(last);
-    }
-    return months;
-  }
-}
-
-/// A class that implements a period type of a semester.
-class SemesterPeriod extends Period implements MonthPeriodBundle {
-  /// A class that implements a period type of a semester.
-  SemesterPeriod({required super.start, required super.end})
-      : assert(
-          end.difference(start) <=
-              const Duration(
-                days: 183,
-                hours: 23,
-                minutes: 59,
-                seconds: 59,
-                milliseconds: 999,
-                microseconds: 999,
-              ),
-          'The difference between start and end must be 183 days, 23 hours, '
-          '59 minutes, 59 seconds, 999 milliseconds and 999 microseconds',
-        ) {
-    if ((start.timeOfDay != Duration.zero) ||
-        (end.timeOfDay != end.endOfDay.timeOfDay) ||
-        start.isAtSameMonthAs(end) ||
-        (start.firstDayOfMonth != start) ||
-        (end.lastDayOfMonth.endOfDay != end) ||
-        ((end.month - start.month) != 5)) {
-      throw ArgumentError.value(
-        end,
-        'end',
-        'End must be at the same semester as start and must be the last '
-            'microsecond of the semester',
-      );
-    }
-  }
-
-  @override
-  List<MonthPeriod> get months {
-    const generator = MonthGenerator();
-    final months = <MonthPeriod>[];
-    var last = generator.of(start);
-    while (last.start.isBefore(end)) {
-      months.add(last);
-      last = generator.after(last);
-    }
-    return months;
-  }
-}
-
-/// A class that implements a period type of a year.
-class YearPeriod extends Period implements MonthPeriodBundle {
-  /// A class that implements a period type of a year.
-  YearPeriod({required super.start, required super.end})
-      : assert(
-          end.difference(start) <=
-              const Duration(
-                days: 365,
-                hours: 23,
-                minutes: 59,
-                seconds: 59,
-                milliseconds: 999,
-                microseconds: 999,
-              ),
-          'The difference between start and end must be 365 days, 23 hours, '
-          '59 minutes, 59 seconds, 999 milliseconds and 999 microseconds',
-        ) {
-    const microsecond = Duration(microseconds: 1);
-    if (!start.isAtSameYearAs(end) ||
-        end.add(microsecond).isAtSameYearAs(start) ||
-        (start.timeOfDay != Duration.zero) ||
-        (end.timeOfDay != end.endOfDay.timeOfDay) ||
-        start.isAtSameMonthAs(end) ||
-        (start.firstDayOfMonth != start) ||
-        (end.lastDayOfMonth.endOfDay != end) ||
-        ((end.month - start.month) != 11)) {
-      throw ArgumentError.value(
-        end,
-        'end',
-        'End must be at the same year as start and must be the last '
-            'microsecond of the year',
-      );
-    }
-  }
-
-  @override
-  List<MonthPeriod> get months {
-    const generator = MonthGenerator();
-    final months = <MonthPeriod>[];
-    var last = generator.of(start);
-    while (last.start.isBefore(end)) {
-      months.add(last);
-      last = generator.after(last);
-    }
-    return months;
-  }
-}
-
-extension _Apply<T extends Object> on T {
-  R apply<R>(R Function(T self) f) => f(this);
 }
